+++ src/Akeneo/Pim/Automation/RuleEngine/front/__tests__/src/factories/index.ts
@@ -1,0 +1,2 @@
+export * from './AttributeFactory';
+export * from './LocaleScopeFactory';