+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/components/Select2Wrapper/Select2Wrapper.tsx
@@ -47,7 +47,7 @@ type Select2GlobalProps = {
   containerCssClass?: string;
   dropdownCssClass?: string;
   hiddenLabel?: boolean;
-  id: string;
+  id?: string;
   label: string;
   onSelecting?: (event: any) => void;
   placeholder?: string;
@@ -59,13 +59,14 @@ type Select2GlobalProps = {
   formatSelection?: (item: Select2Option | Select2OptionGroup) => string;
   hideSearch?: boolean;
   closeTick?: boolean;
+  disabled?: boolean;
 };
 
 type Props = Select2GlobalProps & {
   data?: (Select2Option | Select2OptionGroup)[];
   multiple: boolean;
   ajax?: Select2Ajax;
-  onValueChange?: (value: Select2Value | Select2Value[]) => void;
+  onChange?: (value: Select2Value | Select2Value[]) => void;
   value?: Select2Value | Select2Value[];
 };
 
@@ -84,10 +85,12 @@ const Select2Wrapper: React.FC<Props> = ({
   dropdownCssClass,
   onSelecting,
   value,
-  onValueChange,
+  onChange,
   closeTick = false,
   hideSearch = false,
   allowClear = false,
+  disabled = false,
+  ...remainingProps
 }) => {
   const select2ref = useRef<HTMLInputElement | null>(null);
 
@@ -125,10 +128,11 @@ const Select2Wrapper: React.FC<Props> = ({
         getSelect2Input().on('select2-selecting', onSelecting);
       }
 
-      if (onValueChange) {
+      if (onChange) {
+        getSelect2Input().off('change');
         getSelect2Input().on('change', (e: Select2Event) => {
           const val = e.val;
-          onValueChange(
+          onChange(
             Array.isArray(val) ? (val as Select2Value[]) : (val as Select2Value)
           );
         });
@@ -155,11 +159,21 @@ const Select2Wrapper: React.FC<Props> = ({
 
   useEffect(() => {
     initSelect2(true);
-  }, [JSON.stringify(data)]);
+  }, [onSelecting, allowClear, disabled, JSON.stringify(data)]);
 
   useEffect(() => {
-    initSelect2(true);
-  }, [onSelecting]);
+    if (select2ref.current) {
+      getSelect2Input().off('change');
+      if (onChange) {
+        getSelect2Input().on('change', (e: Select2Event) => {
+          const val = e.val;
+          onChange(
+            Array.isArray(val) ? (val as Select2Value[]) : (val as Select2Value)
+          );
+        });
+      }
+    }
+  }, [onChange]);
 
   useEffect(() => {
     if (select2ref.current) {
@@ -170,7 +184,13 @@ const Select2Wrapper: React.FC<Props> = ({
   return (
     <>
       <Label label={label} hiddenLabel={hiddenLabel} htmlFor={id} />
-      <input id={id} type='hidden' ref={select2ref} />
+      <input
+        id={id}
+        type='hidden'
+        ref={select2ref}
+        disabled={disabled}
+        {...remainingProps}
+      />
     </>
   );
 };