+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/models/actions/SetFamilyAction.ts
@@ -1,36 +1,26 @@
-import React from 'react';
-import { ActionLineProps } from '../../pages/EditRules/components/actions/ActionLineProps';
 import { SetFamilyActionLine } from '../../pages/EditRules/components/actions/SetFamilyActionLine';
 import { FamilyCode } from '../Family';
+import { ActionModuleGuesser } from './ActionModuleGuesser';
 
 export type SetFamilyAction = {
-  module: React.FC<{ action: SetFamilyAction } & ActionLineProps>;
   type: 'set';
   field: 'family';
   value: FamilyCode | null;
 };
 
-export const denormalizeSetFamilyAction = (
-  json: any
-): SetFamilyAction | null => {
+export const getSetFamilyActionModule: ActionModuleGuesser = json => {
   if (json.type !== 'set') {
-    return null;
+    return Promise.resolve(null);
   }
   if (json.field !== 'family') {
-    return null;
+    return Promise.resolve(null);
   }
 
-  return {
-    module: SetFamilyActionLine,
-    type: 'set',
-    field: 'family',
-    value: json.value || null,
-  };
+  return Promise.resolve(SetFamilyActionLine);
 };
 
 export const createSetFamilyAction: () => SetFamilyAction = () => {
   return {
-    module: SetFamilyActionLine,
     type: 'set',
     field: 'family',
     value: null,