+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/pages/EditRules/components/RulesBuilder.tsx
@@ -1,86 +1,50 @@
 import React from 'react';
-import { Router, Translate } from '../../../dependenciesTools';
 import { RuleProductSelection } from './conditions/RuleProductSelection';
-import { RuleDefinition, Locale, LocaleCode } from '../../../models';
-import { Action } from '../../../models/Action';
+import { Condition, Locale, LocaleCode } from '../../../models';
 import { IndexedScopes } from '../../../repositories/ScopeRepository';
 import { ActionLine } from './actions/ActionLine';
-import styled from 'styled-components';
-import middleImage from '../../../assets/illustrations/middle.svg';
-import endImage from '../../../assets/illustrations/end.svg';
-
-const ActionContainer = styled.div`
-  background-image: url('${middleImage}');
-  padding-left: 12px;
-  margin-left: -12px;
-  background-repeat: no-repeat;  
-  padding-bottom: 20px;
-`;
-
-const LastActionContainer = styled.div`
-  background-image: url('${endImage}');
-  padding-left: 12px;
-  margin-left: -12px;
-  background-repeat: no-repeat;  
-`;
+import { Action } from '../../../models/Action';
 
 type Props = {
-  translate: Translate;
-  ruleDefinition: RuleDefinition;
   locales: Locale[];
   scopes: IndexedScopes;
   currentCatalogLocale: LocaleCode;
-  router: Router;
-  actions: (Action | null)[];
   handleDeleteAction: (lineNumber: number) => void;
+  conditions: Condition[];
+  actions: (Action | null)[];
 };
 
 const RulesBuilder: React.FC<Props> = ({
-  translate,
-  ruleDefinition,
   locales,
   scopes,
   currentCatalogLocale,
-  router,
   actions,
   handleDeleteAction,
+  conditions,
 }) => {
-  const isLastAction: (lineNumber: number) => boolean = lineNumber => {
-    const nextActions = actions.slice(lineNumber + 1);
-    return !nextActions.some(action => {
-      return action !== null;
-    });
-  };
-
   return (
     <>
       <RuleProductSelection
-        ruleDefinition={ruleDefinition}
-        translate={translate}
         locales={locales}
         scopes={scopes}
         currentCatalogLocale={currentCatalogLocale}
-        router={router}
+        conditions={conditions}
       />
-      <div data-testid={'action-list'}>
-        {actions.map((action: Action | null, i) => {
-          const Component = isLastAction(i)
-            ? LastActionContainer
-            : ActionContainer;
+      <div data-testid={'action-list'} className={'actionList'}>
+        {actions.map((action, i) => {
           return (
             action && (
-              <Component key={`action_${i}`}>
-                <ActionLine
-                  action={action}
-                  translate={translate}
-                  lineNumber={i}
-                  handleDelete={() => {
-                    handleDeleteAction(i);
-                  }}
-                  router={router}
-                  currentCatalogLocale={currentCatalogLocale}
-                />
-              </Component>
+              <ActionLine
+                action={action}
+                lineNumber={i}
+                handleDelete={() => {
+                  handleDeleteAction(i);
+                }}
+                currentCatalogLocale={currentCatalogLocale}
+                locales={locales}
+                scopes={scopes}
+                key={i}
+              />
             )
           );
         })}