+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/pages/EditRules/components/conditions/RuleProductSelection.tsx
@@ -11,17 +11,20 @@ import {
   createCategoryCondition,
   Locale,
   LocaleCode,
-  RuleDefinition,
 } from '../../../../models/';
 import { TextBoxBlue } from '../TextBoxBlue';
 import { useProductsCount } from '../../hooks';
-import { Router, Translate } from '../../../../dependenciesTools';
 import { IndexedScopes } from '../../../../repositories/ScopeRepository';
 import { ConditionLine } from './ConditionLine';
 import { ProductsCount } from '../ProductsCount';
 import { AddConditionButton } from './AddConditionButton';
 import { FormData } from '../../edit-rules.types';
 import startImage from '../../../../assets/illustrations/start.svg';
+import {
+  useBackboneRouter,
+  useTranslate,
+} from '../../../../dependenciesTools/hooks';
+import { Action } from '../../../../models/Action';
 
 const Header = styled.header`
   font-weight: normal;
@@ -57,67 +60,55 @@ const AddConditionContainer = styled.div`
   padding-left: 15px;
 `;
 
-const getValuesFromFormData = (getValues: Control['getValues']): FormData =>
-  getValues({ nest: true });
-const RuleProductSelectionFieldsetWithAction = styled.fieldset`
-  background-image: url('${startImage}');
-  padding-left: 12px;
-  margin-left: -12px;
-  background-repeat: no-repeat;
+const RuleProductSelectionFieldset = styled.fieldset<{ hasActions: boolean }>`
   padding-bottom: 20px;
+
+  ${({ hasActions }) =>
+    hasActions &&
+    `
+    background-image: url('${startImage}');
+    padding-left: 12px;
+    margin-left: -12px;
+    background-repeat: no-repeat;
+  `}
 `;
 
-const RuleProductSelectionFieldset = styled.fieldset`
-  padding-bottom: 20px;
-`;
+const getValuesFromFormData = (getValues: Control['getValues']): FormData =>
+  getValues({ nest: true });
 
 type Props = {
   currentCatalogLocale: LocaleCode;
   locales: Locale[];
-  router: Router;
-  ruleDefinition: RuleDefinition;
   scopes: IndexedScopes;
-  translate: Translate;
+  conditions: Condition[];
 };
 
 const RuleProductSelection: React.FC<Props> = ({
   currentCatalogLocale,
   locales,
-  router,
-  ruleDefinition,
   scopes,
-  translate,
+  conditions,
 }) => {
-  const [conditions, setConditions] = React.useState<(Condition | null)[]>(
-    ruleDefinition.conditions
-  );
+  const translate = useTranslate();
+  const router = useBackboneRouter();
 
-  const { getValues, unregister, watch } = useFormContext();
-  const deleteCondition = (lineNumber: number) => {
-    Object.keys(getValues()).forEach((value: string) => {
-      if (value.startsWith(`content.conditions[${lineNumber}]`)) {
-        unregister(value);
-      }
-    });
-    setConditions(
-      conditions.map((condition: Condition | null, i: number) => {
-        return i === lineNumber ? null : condition;
-      })
-    );
-  };
+  const [conditionsState, setConditionsState] = React.useState<
+    (Condition | null)[]
+  >([]);
+  React.useEffect(() => {
+    setConditionsState(conditions);
+  }, []);
+
+  const { getValues } = useFormContext();
+
   const productsCount = useProductsCount(
     router,
     getValuesFromFormData(getValues)
   );
-  const watchCondition = (conditionIdentifier: string) => {
-    watch(conditionIdentifier);
-  };
 
-  React.useEffect(() => {
-    setConditions(ruleDefinition.conditions);
-  }, [ruleDefinition]);
-
-  async function createCondition(fieldCode: string): Promise<Condition> {
+  const createCondition: (
+    fieldCode: string
+  ) => Promise<Condition> = async fieldCode => {
     const factories: ConditionFactory[] = [
       createFamilyCondition,
       createTextAttributeCondition,
@@ -134,32 +125,42 @@ const RuleProductSelection: React.FC<Props> = ({
     }
 
     throw new Error(`Unknown factory for field ${fieldCode}`);
-  }
+  };
+
+  const append = (condition: Condition) => {
+    setConditionsState([...conditionsState, condition]);
+  };
+
+  const remove = (lineNumber: number) => {
+    conditionsState[lineNumber] = null;
+    setConditionsState([...conditionsState]);
+  };
 
   const handleAddCondition = (fieldCode: string) => {
-    createCondition(fieldCode).then(condition => {
-      const newConditions = [...conditions, condition];
-      setConditions(newConditions);
-    });
+    createCondition(fieldCode).then(condition => append(condition));
   };
 
-  const isActiveConditionField: (fieldCode: string) => boolean = (
-    fieldCode: string
-  ) => {
-    return conditions.some(condition => {
-      return (
-        condition !== null &&
-        condition.hasOwnProperty('field') &&
-        (condition as { field: string }).field === fieldCode
+  const isActiveConditionField = React.useCallback(
+    (fieldCode: string) => {
+      return (getValues({ nest: true })?.content?.conditions || []).some(
+        (condition: Condition) => {
+          return (
+            Object.hasOwnProperty.call(condition, 'field') &&
+            (condition as { field: string }).field === fieldCode
+          );
+        }
       );
-    });
-  };
+    },
+    [getValues({ nest: true })?.content?.conditions]
+  );
 
-  const Component = ruleDefinition.actions.length
-    ? RuleProductSelectionFieldsetWithAction
-    : RuleProductSelectionFieldset;
+  const hasActions =
+    (getValues({ nest: true })?.content?.actions || []).filter(
+      (action: Action) => action !== null
+    ).length > 0;
+
   return (
-    <Component>
+    <RuleProductSelectionFieldset hasActions={hasActions} tabIndex={-1}>
       <Header className='AknSubsection-title'>
         <HeaderPartContainer>
           <TextBoxBlue>
@@ -173,13 +174,10 @@ const RuleProductSelection: React.FC<Props> = ({
           <ProductsCount
             count={productsCount.value}
             status={productsCount.status}
-            translate={translate}
           />
           <AddConditionContainer>
             <AddConditionButton
-              router={router}
               handleAddCondition={handleAddCondition}
-              translate={translate}
               isActiveConditionField={isActiveConditionField}
             />
           </AddConditionContainer>
@@ -195,20 +193,17 @@ const RuleProductSelection: React.FC<Props> = ({
       </SmallHelper>
       <div className='AknGrid AknGrid--unclickable'>
         <div className='AknGrid-body' data-testid={'condition-list'}>
-          {conditions.map((condition, i) => {
-            watchCondition(`content.conditions[${i}]`);
+          {conditionsState.map((condition, i) => {
             return (
               condition && (
                 <ConditionLine
                   condition={condition}
                   lineNumber={i}
-                  translate={translate}
-                  key={`condition_${i}`}
+                  key={i}
                   locales={locales}
                   scopes={scopes}
                   currentCatalogLocale={currentCatalogLocale}
-                  router={router}
-                  deleteCondition={deleteCondition}
+                  deleteCondition={remove}
                 />
               )
             );
@@ -218,7 +213,7 @@ const RuleProductSelection: React.FC<Props> = ({
       <LegendSrOnly>
         {translate('pimee_catalog_rule.form.legend.product_selection')}
       </LegendSrOnly>
-    </Component>
+    </RuleProductSelectionFieldset>
   );
 };