+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/pages/EditRules/components/conditions/TextAttributeConditionLine.tsx
@@ -7,6 +7,12 @@ import {
 import { ConditionLineProps } from './ConditionLineProps';
 import { InputText } from '../../../../components/Inputs';
 import { AttributeConditionLine } from './AttributeConditionLine';
+import {
+  useBackboneRouter,
+  useTranslate,
+} from '../../../../dependenciesTools/hooks';
+import { Attribute } from '../../../../models/Attribute';
+import { getAttributeByIdentifier } from '../../../../repositories/AttributeRepository';
 
 type TextAttributeConditionLineProps = ConditionLineProps & {
   condition: TextAttributeCondition;
@@ -15,32 +21,36 @@ type TextAttributeConditionLineProps = ConditionLineProps & {
 const TextAttributeConditionLine: React.FC<TextAttributeConditionLineProps> = ({
   condition,
   lineNumber,
-  translate,
   locales,
   scopes,
   currentCatalogLocale,
 }) => {
-  const { register, setValue } = useFormContext();
-
-  const setValueFormValue = (value: string[] | null) =>
-    setValue(`content.conditions[${lineNumber}].value`, value);
+  const router = useBackboneRouter();
+  const translate = useTranslate();
+  const { register } = useFormContext();
+  const [attribute, setAttribute] = React.useState<Attribute | null>();
+  React.useEffect(() => {
+    getAttributeByIdentifier(condition.field, router).then(attribute =>
+      setAttribute(attribute)
+    );
+  });
 
   return (
     <AttributeConditionLine
       condition={condition}
       lineNumber={lineNumber}
-      translate={translate}
       currentCatalogLocale={currentCatalogLocale}
       locales={locales}
       scopes={scopes}
       availableOperators={TextAttributeOperators}
-      setValueFormValue={setValueFormValue}>
+      attribute={attribute}>
       <InputText
         data-testid={`edit-rules-input-${lineNumber}-value`}
         name={`content.conditions[${lineNumber}].value`}
         label={translate('pimee_catalog_rule.rule.value')}
-        ref={register}
+        ref={register()}
         hiddenLabel={true}
+        defaultValue={condition.value}
       />
     </AttributeConditionLine>
   );