+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - PIM-9202: Fix Asset Manager / Product link rules not working with multiple consumers
 - PIM-9261: Fix API assets pagination
 - PIM-9270: Fix assets family product-link-rule definition
+- PIM-9295: Fix error when applying an "Add groups" action to a product model
 
 ## Improvements
 
@@ -30,6 +31,7 @@
 - RUL-20: Rule engine - As Julia, I would like to copy values from/to different attribute types
 - RUL-49: Rule engine - As Peter, I would like to clear attribute values, associations, categories and groups
 - RUL-77: Rule engine - As Peter, I would like to add labels to my rules
+- CLOUD-1959: Use cloud-deployer 2.2 and terraform 0.12.25
 
 ## New features